[Feat] BindingFragment apply - BindingFragment를 통한 ViewBnding관리
