+++ app/src/main/java/org/sopt/sample/presentation/base/BindingFragment.kt
@@ -1,0 +1,32 @@
+package org.sopt.sample.presentation.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+
+abstract class BindingFragment<T : ViewDataBinding>(@LayoutRes private val layoutRes: Int) :
+    Fragment() {
+    private var _binding: T? = null
+    val binding get() = requireNotNull(_binding) { "${this.id} binding error" }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
+        _binding!!.lifecycleOwner = viewLifecycleOwner
+        return binding.root
+    }
+
+
+    override fun onDestroyView() {
+        _binding = null
+        super.onDestroyView()
+    }
+}
