+++ app/src/main/java/org/sopt/sample/presentation/home/fragment/GalleryFragment.kt
@@ -1,35 +1,22 @@
 package org.sopt.sample.presentation.home.fragment
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import org.sopt.sample.R
+import org.sopt.sample.databinding.FragmentGalleryBinding
+import org.sopt.sample.presentation.base.BindingFragment
 
 
-class GalleryFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_gallery, container, false)
-    }
+class GalleryFragment : BindingFragment<FragmentGalleryBinding>(R.layout.fragment_gallery) {
 
     companion object {
 
         @JvmStatic
         fun newInstance() = GalleryFragment()
 
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
     }
 }
